projects/agents/pyproject.toml
[tool.poetry]
name = "agents"
version = "0.1.0"
description = "Agents for Nemesis data processing and interaction."
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
dapr = "^1.15.0"
dapr-ext-fastapi = "^1.15.0"
structlog = "^25.1.0"
pydantic = "^2.10.4"
typing-extensions = "^4.12.2"
gql = {extras = ["requests"], version = "^3.5.3"}
websockets = "^14.2"
# pydantic-ai-slim = {extras = ["retries"], version = "^0.7.1"}
common = { path = "../../libs/common", develop = true }
pydantic-ai = "^0.7.1"
dapr-ext-workflow = "^1.15.0"
logfire = "^4.3.3"
colorlog = "^6.9.0"
opentelemetry-instrumentation-fastapi = "^0.57b0"
opentelemetry-api = "^1.36.0"
opentelemetry-sdk = "^1.36.0"
opentelemetry-instrumentation = "^0.57b0"
opentelemetry-exporter-otlp-proto-grpc = "^1.36.0"
opentelemetry-exporter-otlp-proto-http = "^1.36.0"
openinference-instrumentation-pydantic-ai = {version = "^0.1.4", python = ">=3.12,<3.14"}
asyncpg = "^0.30.0"
psycopg = "^3.2.9"
pythonnet = "^3.0.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

